Update README.md
1. You have multiple print statements asking for maths percentage.

2. UCT specifies that any results below 40% do not attract score. Make use of if statements to account for instances where a students score for a subject is <40.

3. UCT has faculty specific requirements. This line is taken from the brochure: "...but including any other required subject(s) for the relevant 
programme. (For a programme where Mathematics and Physical Sciences are required, take the 
scores for English, Mathematics, Physical Sciences and the next best three subjects other than Life Orientation and AP 
subjects to calculate your APS.)".

Example: The faculty of business science has the below admission requirements as the minimum requirements for admission:

Possible admission into Education 
Development Unit only 
FPS of 430 to 434.
NBT scores of Upper Intermediate for AL & QL  //Dont pay attention to this requirement
Mathematics 60% 
English HL 50% 
English FAL 60%

Only add conditions for the minimum requirements. So if the student meets the minimum requirements as stated above, we can let the user know that they are eligible to apply for this faculty. This applies for all other faculties.